chore: contract formatting with forge

+++ src/Counter.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.13;
 
 contract Counter {
-    uint public number;
+    uint256 public number;
 
-    function get() public view returns (uint) {
+    function get() public view returns (uint256) {
         return number;
     }
